+++ tests/Type/EnumTypeTest.php
@@ -551,8 +551,8 @@ class EnumTypeTest extends TestCase
             ],
             'errors' => [
                 [
-                    'debugMessage' => 'Expected a value of type Complex but received: instance of ArrayObject. Cannot serialize value as enum: instance of ArrayObject',
                     'locations'    => [['line' => 5, 'column' => 9]],
+                    'extensions' => ['debugMessage' => 'Expected a value of type Complex but received: instance of ArrayObject. Cannot serialize value as enum: instance of ArrayObject'],
                 ],
             ],
         ];
@@ -606,10 +606,12 @@ class EnumTypeTest extends TestCase
                 'data'   => ['first' => 'ONE', 'second' => 'TWO', 'third' => null],
                 'errors' => [
                     [
-                        'debugMessage' => 'Expected a value of type SimpleEnum but received: WRONG. Cannot serialize value as enum: WRONG',
                         'locations'    => [['line' => 4, 'column' => 13]],
-                        'trace' => [
-                            ['call' => 'GraphQL\Type\Definition\EnumType::serialize()'],
+                        'extensions' => [
+                            'debugMessage' => 'Expected a value of type SimpleEnum but received: WRONG. Cannot serialize value as enum: WRONG',
+                            'trace' => [
+                                ['call' => 'GraphQL\Type\Definition\EnumType::serialize()'],
+                            ],
                         ],
                     ],
                 ],